requirements.txt
github3.py
PyYAML
